lib/generators/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-jsonpath>=1.3.1",
    "pytypid",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.4.1",
]

[tool.uv.sources]
pytypid = { git = "https://github.com/kit-data-manager/typid-client.git" }
